+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [5.1.0](https://github.com/Space48/cloud-seed/compare/v5.0.2...v5.1.0)
+
+### Added
+
+- Added the ability to connect gen2 functions to a VPC using direct VPC egress
+- Added the ability to connect gen1 and gen2 functions to a VPC using a VPC connector
+
 ## [5.0.2](https://github.com/Space48/cloud-seed/compare/v5.0.1...v5.0.2)
 
 ### Fixed